tg_digest_system/tg_digest_system/web/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
psycopg2-binary==2.9.9
pydantic==2.5.0
telethon==1.34.0
aiohttp==3.9.1
httpx>=0.27.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
